memccpy test

+++ ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: pgernez <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/06 23:23:14 by pgernez           #+#    #+#             */
-/*   Updated: 2017/01/07 23:09:08 by pgernez          ###   ########.fr       */
+/*   Updated: 2017/02/15 21:13:48 by pgernez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@ void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
 	size_t	i;
 
 	i = 0;
-	while (i != c && i <= (n - 1))
+	while ((int)i != c && i <= (n - 1))
 	{
 		((char*)dst)[i] = ((char*)src)[i];
 		i++;
 	}
-	if (i == c)
+	if ((int)i == c)
 	{
 		((char*)dst)[i] = ((char*)src)[i];
 		i++;
